==> Notebooks/2d - Project Buurtscores - XGBoost.ipynb <==
# Model in XGBoost maken

Voor oordelen van zowel Erwin als Thomas, met alle features en zonder woz vanwege zeer veel missing data daarin terwijl het wel een belangrijke factor is.


#### Input
- Scores voor 300+ buurten door Erwin (dus aangevuld met meer b/c-buurten voor balans in dataset)
- Buurten.csv bestand met alle kolommen

#### Output
- Modellen
- Bestanden met predicties op alle buurten

## Korte uitleg werking modellen


#### Decision tree
Je bepaalt een mate van 'wanorde' en laat hem de data splitten op basis van businessregels zodat in elke stap de mate van wanorde zo klein mogelijk wordt. Hiervoor zijn onder de motorkap verschillende algoritmes mogelijk. <br><br>
Probleem is dat deze snel dingen 'leert' die eigenlijk niet relevant zijn maar toevallig in de dataset waarop het getraind is zitten, maar niet gelden voor toekomstige data. Hierdoor wordt het inaccuraat en dit heet overfitting. Dit model is volledig weer te geven en te interpreteren.

#### Bagging (bootstrap aggregating)
Overfitting kan gereduceerd worden door niet 1 boom, maar paralel veel bomen te maken op willekeurig gekozen onderdelen van de dataset, waarvan je dan het gemiddelde neemt. Hierdoor zullen niet echt bestaande effecten die voor overfitting zorgen minder belangrijk worden en wordt het model accurater.

#### Random forest
Deze gebruikt bagging, maar laat het algoritme per boom ook maar een percentage van de features (dus niet alleen de rijen) gebruiken, dit ook om overfitting te voorkomen. Deze geeft vaak goede resultaten en is vaak simpel te optimaliseren.

####  Boosting
Deze bouwt een (simpele) boom per keer, en gaat per keer de fouten van de vorige boom proberen te corrigeren. Vervolgens worden deze gewogen en zorgen die voor een betere voorspeller. Deze kan verbeterd worden door het gradient descent algoritme te gebruiken om de optimale 'fit' te vinden. Deze gebruikt de gradient (helling) van de loss-functie (de functie die zegt wat er fout is, bijv: (echte waarde - voorspelde waarde)^2  ) daarvoor.

#### XGBoost
Dit is een versie van gradient descend boosting die in de back-end geoptimaliseerd is voor optimalisatie, gebruiksgemak, en efficiënt omgaan met computerkracht. 



# importeren packages
import pandas as pd
import numpy as np
#conda install -c anaconda py-xgboost #installeer XGBoost
import xgboost
#from xgboost import XGBClassifier
from xgboost import plot_importance
from sklearn.model_selection import train_test_split, RandomizedSearchCV, GridSearchCV, KFold, cross_val_score
from sklearn.metrics import accuracy_score, confusion_matrix
from sklearn import preprocessing
from matplotlib import pyplot
#import seaborn as sns #visualisation
#import matplotlib.pyplot as plt #visualisation

# setten source path: later moet alle brondata naar datalake
source_path ="P:\\DATASCIENCE\\Aantrekkelijkheid buurten\\Resultaten\\"
# alle tussentijdse data voorlopig in dezelfde map op een schijf opslaan in
save_path = "P:\\DATASCIENCE\\Aantrekkelijkheid buurten\\Resultaten\\"


#%matplotlib inline

# display meer regels
pd.set_option('display.max_rows', 500)

### Inladen data

# inladen bestand met scores als sc
sc = pd.read_excel(source_path + 'buurten voor experts 2e ronde.xlsx', sheet_name='buurten')
# alleen selecteren kolommen voor scores en oordelen
sc = sc[['buurtcode','EE','TK']]

# inladen bestand met alle buurten en onafhankelijke variabelen als all
buurtendata = pd.read_csv(source_path + 'buurten.csv', encoding='utf-16')

# merge op buurtcode (die van 2019)
df = pd.merge(sc,buurtendata, on='buurtcode')

### Maak splits in test en training set

# alle kolommen! behalve enkele omschhrijvingen, onnodige wijkcodes, marktpotentie want String 
feature_cols = ['woxdelta5jrGem', 'AantalCafe3', 'woxdelta5jr', 'ar60', 'wox',
       'woxdelta1jrGem', 'AantalBios5', 'woxm2WG']




# ['AantalCafe3', 'ar60', 'woxdelta5jr', 'woxm2', 'woxww',
#        'woxdelta5jrGem', 'wox', 'RLBBEV18', 'ov60',
#        'RLBVRZ16', 'AantalLevensm3', 'woxm2WG', 'aand1945-1970',
#        'woxdelta5jrWGabs', 'woxdelta1jrGem', 'fiets', 'AantalPodiumk10',
#        'banen', 'wv2000-2010', 'AantalBios5']




#zonder wox
#feature_cols = ['buurtcode','inw', 'bevdh', 'oadh', 'hh', 'inw65', 'KL16', 'KL18', 'VKL1618',
#                'RLBRMTR18', 'RLBWON18', 'RLBBEV18', 'RLBVRZ18', 'RLBVEI18', 'RLBFYS18', 
#                'AfstandSupermarkt', 'AantalLevensm3', 'AfstandHuisarts', 'AantalCafe3','AantalRest3', 'AantalKinderopvang1', 
#                'AfstandZwembad', 'AantalMusea10', 'AantalPodiumk10', 'AantalBios5', 
#                'hh19inkLaag', 'hh30inkLaag', 'hh19inkMidden', 'hh30inkMidden', 'hh19inkHoog', 'hh30inkHoog', 
#                'aandeelMH19', 'auto30', 'auto60', 'ov30', 'ov60', 'fiets', 'banen'] 



X = df[feature_cols]
y = df[['TK']]

## hyperparameter search

- 3*10-fold cross-validation met random seeds
- en holdout set om te verzekeren dat je niet je cross-validatie overfit

- gekozen om gewoon bomen te gebruiken als booster
- model in principe niet exact reproducible ook als we random state
- base_score zou lager kunnen omdat we voor 3 groepen classificeren, maar we geven hem genoeg stappen om te trainen dus deze maakt niet echt uit




# Laad model
clf = xgboost.XGBClassifier(seed=2)

clf.__dict__

# test hyperparameter search input (per parameter terugbrengen naar minder waarden)
#    "max_depth"        : [2, 3, 4, 5, 6, 7, 9, 12, 19, 25],
#    "learning_rate"    : [0.01, 0.05, 0.10, 0.15, 0.20, 0.30, 0.50, 0.80],
#    "subsample"        : [0.1, 0.4, 0.6, 0.8, 1],
#    "min_child_weight" : [1, 3, 5, 7],
#    "gamma"            : [0.0, 0.1, 0.2, 0.3, 0.4],
#    "colsample_bytree" : [0.3, 0.4, 0.5, 0.7, 1],
#    "n_estimators"     : [10, 100, 250, 500]



# hyperparameter search input
params = {
    "max_depth"        : [2, 5, 7, 9],
    "learning_rate"    : [0.10, 0.15, 0.20],
    "subsample"        : [0.8, 1],
    "min_child_weight" : [3, 5, 7],
    "gamma"            : [0.1, 0.2, 0.3],
    "colsample_bytree" : [0.4, 0.5, 0.7],
    "n_estimators"     : [100, 250]
}

grid_search = GridSearchCV(
    estimator = clf,
    param_grid = params,
    scoring = 'f1_macro',
    n_jobs = 10,
    cv = 10,
    verbose = True
)

grid_search.fit(X, y)

grid_search.best_estimator_

# splitten
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=2)

# voor geselecteerde kolommen, fase 2, TK

XGBClassifier(base_score=0.5, booster='gbtree', colsample_bylevel=1,
              colsample_bynode=1, colsample_bytree=0.7, gamma=0.2,
              learning_rate=0.15, max_delta_step=0, max_depth=7,
              min_child_weight=3, missing=None, n_estimators=250, n_jobs=1,
              nthread=None, objective='multi:softprob', random_state=0,
              reg_alpha=0, reg_lambda=1, scale_pos_weight=1, seed=2,
              silent=None, subsample=0.8, verbosity=1)

#### voor alle kolommen - EE


XGBClassifier(base_score=0.5, booster='gbtree', colsample_bylevel=1,
              colsample_bynode=1, colsample_bytree=0.4, gamma=0.1,
              learning_rate=0.15, max_delta_step=0, max_depth=7,
              min_child_weight=5, missing=None, n_estimators=100, n_jobs=1,
              nthread=None, objective='multi:softprob', random_state=2,
              reg_alpha=0, reg_lambda=1, scale_pos_weight=1, seed=2,
              silent=None, subsample=0.8, verbosity=1)

#### voor zonder wox - EE


XGBClassifier(base_score=0.5, booster='gbtree', colsample_bylevel=1,
              colsample_bynode=1, colsample_bytree=0.4, gamma=0.3,
              learning_rate=0.15, max_delta_step=0, max_depth=7,
              min_child_weight=3, missing=None, n_estimators=100, n_jobs=1,
              nthread=None, objective='multi:softprob', random_state=2,
              reg_alpha=0, reg_lambda=1, scale_pos_weight=1, seed=2,
              silent=None, subsample=0.8, verbosity=1)

#### voor alle kolommen - TK


XGBClassifier(base_score=0.5, booster='gbtree', colsample_bylevel=1,
              colsample_bynode=1, colsample_bytree=0.4, gamma=0.1,
              learning_rate=0.15, max_delta_step=0, max_depth=7,
              min_child_weight=3, missing=None, n_estimators=100, n_jobs=1,
              nthread=None, objective='multi:softprob', random_state=2,
              reg_alpha=0, reg_lambda=1, scale_pos_weight=1, seed=2,
              silent=None, subsample=1, verbosity=1)

#### voor zonder wox - TK


XGBClassifier(base_score=0.5, booster='gbtree', colsample_bylevel=1,
              colsample_bynode=1, colsample_bytree=0.7, gamma=0.3,
              learning_rate=0.15, max_delta_step=0, max_depth=9,
              min_child_weight=3, missing=None, n_estimators=100, n_jobs=1,
              nthread=None, objective='multi:softprob', random_state=2,
              reg_alpha=0, reg_lambda=1, scale_pos_weight=1, seed=2,
              silent=None, subsample=1, verbosity=1)

# maak object met geoptimaliseerde hyperparameters
model = xgboost.XGBClassifier(base_score=0.5, booster='gbtree', colsample_bylevel=1,
              colsample_bynode=1, colsample_bytree=0.7, gamma=0.2,
              learning_rate=0.15, max_delta_step=0, max_depth=7,
              min_child_weight=3, missing=None, n_estimators=250, n_jobs=1,
              nthread=None, objective='multi:softprob', random_state=0,
              reg_alpha=0, reg_lambda=1, scale_pos_weight=1, seed=2,
              silent=None, subsample=0.8, verbosity=1)

# en fit op data
model.fit(X_train, y_train)

y_pred = model.predict(X_test)
y_train_pred = model.predict(X_train)

train_accuracy = accuracy_score(y_train,y_train_pred)
test_accuracy = accuracy_score(y_test,y_pred)
cm = confusion_matrix(y_test, y_pred)

train_accuracy

test_accuracy

cm

model.feature_importances_

# maak dataframe met feature importances
fi = pd.DataFrame(model.feature_importances_,X_train.columns)
fi.columns = ['feature_score']
fi.sort_values('feature_score', ascending=False, inplace=True)
fi['feature'] = fi.index
fi

len(feature_cols)

fi[fi['feature_score']>0.05]['feature'].values

#fi.to_csv('fi EE zonder wox.csv', encoding='utf-16', index=False)

### Predicten rest van buurten en maken bestand

prb = buurtendata[feature_cols]

pred = model.predict(prb)
pred = pd.DataFrame(pred,columns=['pred_TK'])
prb['pred_TK'] = pred

prb

#prb.to_csv('predicties TK alle data, zonder wox.csv', encoding='utf-16', index=False)
