==> Notebooks/3c - Project Buurtscores - runnen backtesten.ipynb <==
# Backtest uitvoeren

Op 2020 data


# importeren packages
import pandas as pd
import numpy as np
#conda install -c anaconda py-xgboost #installeer XGBoost
import xgboost
#from xgboost import XGBClassifier
from xgboost import plot_importance
from sklearn.model_selection import train_test_split, RandomizedSearchCV, GridSearchCV, KFold, cross_val_score
from sklearn.metrics import accuracy_score, confusion_matrix
from sklearn import preprocessing
from sklearn.preprocessing import RobustScaler 
import pickle
from matplotlib import pyplot
#import seaborn as sns #visualisation
#import matplotlib.pyplot as plt #visualisation

# setten source path: later moet alle brondata naar datalake
source_path ="P:\\DATASCIENCE\\Aantrekkelijkheid buurten\\Resultaten\\"
# alle tussentijdse data voorlopig in dezelfde map op een schijf opslaan in
save_path = "P:\\DATASCIENCE\\Aantrekkelijkheid buurten\\Resultaten\\"


#%matplotlib inline

# display meer regels
pd.set_option('display.max_rows', 500)

### Inladen data

# # inladen bestand met scores als sc
# sc = pd.read_excel(source_path + 'buurten voor experts 2e ronde.xlsx', sheet_name='buurten')
# # alleen selecteren kolommen voor scores en oordelen
# sc = sc[['buurtcode','EE','TK']]

# inladen bestand met alle buurten en onafhankelijke variabelen als all
buurtendata = pd.read_csv(source_path + '2017\\' +  'buurten.csv', encoding='utf-16')

# merge op buurtcode (die van 2019)
# df = pd.merge(sc,buurtendata, on='buurtcode')
df = buurtendata

### Maak splits in test en training set

df.columns

# hernoem RLBBEV naar zonder jaartal, zodat we kunnen vergelijken tussen 2016 en 2018
df.rename(columns={"RLBBEV16":"RLBBEV"}, inplace=True) 

# alle kolommen! behalve enkele omschhrijvingen, onnodige wijkcodes, marktpotentie want String 
# feature_cols = ['woxdelta5jrGem', 'AantalCafe3', 'woxdelta5jr', 'wox',
#        'woxdelta1jrGem', 'AantalBios5', 'woxm2WG']




feature_cols = ['AantalCafe3', 'woxdelta5jr', 'woxm2', 'woxww',
       'woxdelta5jrGem', 'wox', 'RLBBEV',
       'RLBVRZ16', 'AantalLevensm3', 'woxm2WG', 'aand1945-1970',
       'woxdelta5jrWGabs', 'woxdelta1jrGem', 'AantalPodiumk10',
       'wv2000-2010', 'AantalBios5']




#zonder wox
#feature_cols = ['buurtcode','inw', 'bevdh', 'oadh', 'hh', 'inw65', 'KL16', 'KL18', 'VKL1618',
#                'RLBRMTR18', 'RLBWON18', 'RLBBEV18', 'RLBVRZ18', 'RLBVEI18', 'RLBFYS18', 
#                'AfstandSupermarkt', 'AantalLevensm3', 'AfstandHuisarts', 'AantalCafe3','AantalRest3', 'AantalKinderopvang1', 
#                'AfstandZwembad', 'AantalMusea10', 'AantalPodiumk10', 'AantalBios5', 
#                'hh19inkLaag', 'hh30inkLaag', 'hh19inkMidden', 'hh30inkMidden', 'hh19inkHoog', 'hh30inkHoog', 
#                'aandeelMH19', 'auto30', 'auto60', 'ov30', 'ov60', 'fiets', 'banen'] 



#X = df[feature_cols]
# y = df[['TK']]

# Normaliseren data
# Creating a scaled feature set
scaler = RobustScaler()
X_2017 = scaler.fit_transform(df[feature_cols])
X_2017 = pd.DataFrame(X_2017)
X_2017.columns = feature_cols

model_2019 = xgboost.XGBClassifier(base_score=0.5, booster='gbtree', colsample_bylevel=1,
              colsample_bynode=1, colsample_bytree=0.7, gamma=0.2,
              learning_rate=0.15, max_delta_step=0, max_depth=7,
              min_child_weight=3, missing=None, n_estimators=250, n_jobs=1,
              nthread=None, objective='multi:softprob', random_state=0,
              reg_alpha=0, reg_lambda=1, scale_pos_weight=1, seed=2,
              silent=None, subsample=0.8, verbosity=1)

# laad model
model_2019 = pickle.load(open("model_backtesting.pkl", "rb"))


y_pred = model_2019.predict(X_2017)

y_pred.shape

X_2017.columns

y_pred

# selecteer alleen kolommen om op te vergelijken en key, absolute getallen
X_oud = df[["buurtcode", "woxww"]]
# voeg predicties toe
X_oud['pred_2017'] = list(y_pred)

# laad 2019 data in voor vergelijking
X_nieuw = pd.read_csv(source_path + 'buurten.csv', encoding='utf-16')
# selecteer alleen kolommen om op te vergelijken
X_nieuw = X_nieuw[['buurtcode', 'woxww', 'VKL1618']]

# mergen beiden
X_vgl = pd.merge(X_oud, X_nieuw, on='buurtcode')

# berekenen verschil woxww
X_vgl['delta_woxww'] = X_vgl['woxww_y'] - X_vgl['woxww_x']

X_vgl = X_vgl[X_vgl['delta_woxww'].notnull()]

X_vgl.shape

X_vgl.groupby('pred_2017')['delta_woxww'].mean()

X_vgl.groupby('pred_2017')['delta_woxww'].count()


X_vgl = X_vgl[X_vgl['VKL1618'].notnull()]
X_vgl.groupby('pred_2017')['VKL1618'].mean()
