==> Notebooks/3d - Project Buurtscores - Vergelijken Conf Modellen.ipynb <==
# Comparing Normalized & Conf Models
Via XGB.train kunnen wij aparte confidence berekenen
via xgbclassifier.fit niet

# importeren packages
import pandas as pd
import numpy as np
#conda install -c anaconda py-xgboost #installeer XGBoost
import xgboost
from xgboost import XGBClassifier
from xgboost import plot_importance
from sklearn.model_selection import train_test_split, RandomizedSearchCV, GridSearchCV, KFold, cross_val_score 
from sklearn.metrics import accuracy_score, confusion_matrix
from matplotlib import pyplot
from sklearn.preprocessing import RobustScaler

# inladen bestand met scores als sc
sc = pd.read_excel('P:\\DATASCIENCE\\Aantrekkelijkheid buurten\\Resultaten\\buurten voor experts 2e ronde.xlsx', sheet_name='buurten')
# alleen selecteren kolommen voor scores en oordelen
sc = sc[['buurtcode','EE','TK']]

# inladen bestand met alle buurten en onafhankelijke variabelen als all
buurtendata = pd.read_csv('P:\\DATASCIENCE\\Aantrekkelijkheid buurten\\Resultaten\\buurten.csv', encoding='utf-16')

df = pd.merge(sc,buurtendata, on='buurtcode')

# feature_cols_full = ['inw', 'bevdh', 'oadh', 'hh', 'inw65', 'KL16', 'KL18', 'VKL1618',
#                'RLBRMTR18', 'RLBWON18', 'RLBBEV18', 'RLBVRZ18', 'RLBVEI18', 'RLBFYS18', 
#                'AfstandSupermarkt', 'AantalLevensm3', 'AfstandHuisarts', 'AantalCafe3','AantalRest3', 'AantalKinderopvang1', 
#                'AfstandZwembad', 'AantalMusea10', 'AantalPodiumk10', 'AantalBios5', 
#                'hh19inkLaag', 'hh30inkLaag', 'hh19inkMidden', 'hh30inkMidden', 'hh19inkHoog', 'hh30inkHoog', 
#                'aandeelMH19', 'auto30', 'auto60', 'ov30', 'ov60', 'fiets', 
#                'wox', 'woxdelta1jr', 'woxdelta5jr', 'woxww', 'woxm2', 
#                'woxGem', 'woxdelta1jrGem', 'woxdelta5jrGem', 'woxwwGem', 'woxm2Gem', 'woxdelta5jrWGabs', 'woxdelta5jrWGrel', 
#                'woxm2WG','banen'] 

# feature_cols_full = ['AantalCafe3', 'ar60', 'woxdelta5jr', 'woxm2', 'woxww',
#        'woxdelta5jrGem', 'wox', 'RLBBEV18', 'ov60',
#        'RLBVRZ16', 'AantalLevensm3', 'woxm2WG', 'aand1945-1970',
#        'woxdelta5jrWGabs', 'woxdelta1jrGem', 'fiets', 'AantalPodiumk10',
#        'banen', 'wv2000-2010', 'AantalBios5']

feature_cols_full = ['woxdelta5jrGem', 'AantalCafe3', 'woxdelta5jr', 'ar60', 'wox',
       'woxdelta1jrGem', 'AantalBios5', 'woxm2WG']

X = df[feature_cols_full]

# Same for all models
replace_dict = {'a': 0, 'b': 1, 'c':2}
y = df['TK'].apply(lambda x: replace_dict[x])

# Creating a scaled feature set
scaler = RobustScaler()
X_scaled = scaler.fit_transform(df[feature_cols_full])
X_scaled = pd.DataFrame(X_scaled)
X_scaled.columns = feature_cols_full

#X_scaled.columns

# Same for all models
# replace_dict = {'a': 0, 'b': 1, 'c':2}
# y = df['EE'].apply(lambda x: replace_dict[x])

## Testing
#pd.DataFrame({'original': df['EE'], 'new': y}).to_csv('testing.csv')

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=2)
X_train_scaled, X_test_scaled, y_train_scaled, y_test_scaled = train_test_split(X_scaled, y, test_size=0.25, random_state=2)

# Test if split is the same indices are taken
print(sum(y_train_scaled.index.values == y_train.index.values)/len(y_train))
print(sum(X_train_scaled.index.values == X_train.index.values)/len(y_train))

# check if < in column names, this is not allowed
#[x for x in X_train.columns if '<' in x]

# For the confidence model we factorize the output
# replace_dict = {'a': 0, 'b': 1, 'c':2}

# y_train = y_train['EE'].apply(lambda x: replace_dict[x])
# y_test = y_test['EE'].apply(lambda x: replace_dict[x])

# y_train_scaled = y_train_scaled['EE'].apply(lambda x: replace_dict[x])
# y_test_scaled = y_test_scaled['EE'].apply(lambda x: replace_dict[x])

# Testing

## Creating models with confidence for both datasets; standard & scaled
### Conf model with standard data

params= {'base_score':0.5, 'booster':'gbtree', 'colsample_bylevel':1,
              'colsample_bynode':1, 'colsample_bytree':0.7, 'gamma':0.2,
              'learning_rate':0.15, 'max_delta_step':0, 'max_depth':7,
              'min_child_weight':3, 'missing':None, 'n_estimators':250, 'n_jobs':1,
              'nthread':None, 'objective':'multi:softprob', 'random_state':0,
              'reg_alpha':0, 'reg_lambda':1, 'scale_pos_weight':1, 'seed':2,
              'silent':None, 'subsample':0.8, 'verbosity':3, 'num_class' : 3}

xgb_train = xgboost.DMatrix(X_train, label=y_train)
xgb_test =  xgboost.DMatrix(X_test, label=y_test)
model = xgboost.train(params, dtrain = xgb_train)

test_preds = [x.argmax() for x in model.predict(xgb_test)]

for x,y,z in zip(y_test, test_preds, [x.max() for x in model.predict(xgb_test)]):
    print(f"Truth: {x}, pred:{y}, conf {z}")

#model.predict(xgb_test)

model.save_config()

#train_accuracy = accuracy_score(y_train,test_preds)
test_accuracy = accuracy_score(y_test,test_preds)
cm = confusion_matrix(y_test, test_preds)

test_accuracy

model.get_score(importance_type='gain')

standard_conf_preds = test_preds

### Conf model with scaled data

X_train[:1]

X_train_scaled[:1]

y_train_scaled[:1]

y_train[:1]

xgb_train_scaled = xgboost.DMatrix(X_train_scaled, label=y_train_scaled)
xgb_test_scaled =  xgboost.DMatrix(X_test_scaled, label=y_test_scaled)
model = xgboost.train(params, dtrain = xgb_train_scaled)

test_preds = [x.argmax() for x in model.predict(xgb_test_scaled)]

for x,y,z in zip(y_test_scaled, test_preds, [x.max() for x in model.predict(xgb_test_scaled)]):
    print(f"Truth: {x}, pred:{y}, conf {z}")

# confidence verifieren. 

normalized_conf_preds = test_preds

# Standard Model standard data

# maak object met geoptimaliseerde hyperparameters
standard_model = xgboost.XGBClassifier(base_score=0.5, booster='gbtree', colsample_bylevel=1,
              colsample_bynode=1, colsample_bytree=0.7, gamma=0.2,
              learning_rate=0.15, max_delta_step=0, max_depth=7,
              min_child_weight=3, missing=None, n_estimators=250, n_jobs=1,
              nthread=None, objective='multi:softprob',softmax_multiclass_param= 3,random_state=0,
              reg_alpha=0, reg_lambda=1, scale_pos_weight=1, seed=2,
              silent=None, subsample=0.8, verbosity=1)

# en fit op data
standard_model.fit(X_train, y_train)

#standard_model.get_xgb_params

y_pred = standard_model.predict(X_test)
y_train_pred = standard_model.predict(X_train)

train_accuracy = accuracy_score(y_train,y_train_pred)
test_accuracy = accuracy_score(y_test,y_pred)
cm = confusion_matrix(y_test, y_pred)
test_accuracy

standard_standard_preds = y_pred

standard_standard_preds

y_train

# fi = pd.DataFrame(standard_model.feature_importances_,X_train.columns)
# fi.columns = ['feature_score']
# fi.sort_values('feature_score', ascending=False, inplace=True)
# fi['feature'] = fi.index
# fi

## Standard Model Normalized Data

standard_model = xgboost.XGBClassifier(base_score=0.5, booster='gbtree', colsample_bylevel=1,
              colsample_bynode=1, colsample_bytree=0.7, gamma=0.2,
              learning_rate=0.15, max_delta_step=0, max_depth=7,
              min_child_weight=3, missing=None, n_estimators=250, n_jobs=1,
              nthread=None, objective='multi:softprob', random_state=0,
              reg_alpha=0, reg_lambda=1, scale_pos_weight=1, seed=2,
              silent=None, subsample=0.8, verbosity=1)

standard_model.fit(X_train_scaled, y_train_scaled)

standard_model.get_params

y_pred = standard_model.predict(X_test_scaled)
y_train_pred = standard_model.predict(X_train_scaled)

train_accuracy = accuracy_score(y_train_scaled,y_train_pred)
test_accuracy = accuracy_score(y_test_scaled,y_pred)
cm = confusion_matrix(y_test, y_pred)
test_accuracy

cm

standard_normalized_preds = y_pred

# Comparing Results

results = pd.DataFrame({
    'conf_standard': standard_conf_preds,
    'conf_normalized': normalized_conf_preds,
    'standard_standard': standard_standard_preds,
    'standard_normalized': standard_normalized_preds
})
#     standard_normalized_preds
# standard_standard_preds
# normalized_conf_preds
# standard_conf_preds
results.head()

#pd.Series(results.sum(axis=1)%4)

#pd.Series(results.sum(axis=1)%4).sum()

results.corr()

# Zijn dus een paar verschillen

standard_model.save_model('0001.model') 

bst = xgboost.Booster(params)  # init model
bst.load_model('0001.model')  # load data

bst = xgboost.XGBClassifier(base_score=0.5, booster='gbtree', colsample_bylevel=1,
              colsample_bynode=1, colsample_bytree=0.7, gamma=0.2,
              learning_rate=0.15, max_delta_step=0, max_depth=7,
              min_child_weight=3, missing=None, n_estimators=250, n_jobs=1,
              nthread=None, objective='multi:softprob', random_state=0,
              reg_alpha=0, reg_lambda=1, scale_pos_weight=1, seed=2,
              silent=None, subsample=0.8, verbosity=1)
bst.load_model('0001.model') 

bst
