==> Notebooks/4d - voorspellingen en verklaringen combineren.ipynb <==
# De Finale Modellen Creeren en hier een verklaring bij geven.
Hierbij zijn voor Thomas en Erwin 2 modellen gemaakt: met en zonder WOX.


Indien WOX data beschikbaar is, gebruiken wij het wox model. Hier wordt uitgegaan van een optimaal iteraties at 75. Voor het niet WOX model wordt een aantal iteraties van X gebruikt. 
Deze waardes zijn om overfitting te voorkomen. Wij willen het model alleen dingen leren die te generaliseren zijn. 


De WOX en niet WOX modellen worden samengevoegd om tot een Ensemble te komen.

Via SHAP values worden verklaringen gegeven voor waarom een voorspelling wordt gedaan. 

Per buurt willen wij hierdoor per emsemble model (= wox/geen wox samengevoegd)
* Voorspelling TK 
* Voorspelling EE 
* Confidence TK voorspellingen 
* Confidence EE voorspellingen
* Feature Importance TK
* Feature Importance EE

---------------
In deze notebook realiseren wij een output voor Powerbi
- samenvoegen van de 4 modellen tot 2 waarbij wox/ geen wox wordt gecombineerd.
- confidence toevoegen
- 5 belangrijkste features meegeven.

import pandas as pd

predictions = pd.read_csv('../2019model_predictions.csv')

predictions.head()

# setten source path: later moet alle brondata naar datalake
source_path ="P:\\DATASCIENCE\\Aantrekkelijkheid buurten\\Resultaten\\"
# alle tussentijdse data voorlopig in dezelfde map op een schijf opslaan in
save_path = "P:\\DATASCIENCE\\Aantrekkelijkheid buurten\\Resultaten\\"

# inladen bestand met alle buurten en onafhankelijke variabelen als all
buurtendata = pd.read_csv(source_path + 'buurten.csv', encoding='utf-16')
buurtendata.rename(columns={"RLBBEV18":"RLBBEV",'KL18': "KL", 
               'RLBRMTR18': 'RLBRMTR', 'RLBWON18':'RLBWON' , 'RLBBEV18':'RLBBEV', 'RLBVRZ18':'RLBVRZ', 'RLBVEI18': 'RLBVEI', 'RLBFYS18': 'RLBFYS',}, inplace=True) 

predictions['wox_info'] = buurtendata['wox'].notnull()
predictions['EE_Ensemble'] = predictions.apply(lambda x: x['ee_wox'] if x['wox_info'] else x['ee_nowox'], axis=1)
predictions['TK_Ensemble'] = predictions.apply(lambda x: x['TK_wox'] if x['wox_info'] else x['TK_nowox'], axis=1)


predictions.head()

## Identifier van het model gaan wij gebruiken om de bijbehorende informatie op te halen
wox -> wox uitleg

ee_nowox = pd.read_csv('../interpretations/final_model_EE_nowox_interpretations.csv')
ee_wox = pd.read_csv('../interpretations/final_model_EE_wox_interpretations.csv')
TK_nowox = pd.read_csv('../interpretations/final_model_TK_nowox_interpretations.csv')
TK_wox = pd.read_csv('../interpretations/final_model_TK_wox_interpretations.csv')

ee_nowox.head()

def capture_conf_ee(x):
    # check index
    index = x.ix[0]
    if x['wox_info']:
        return ee_wox.iloc[index]['confidence']
    else:
        return ee_nowox.iloc[index]['confidence']
capture_conf(predictions.iloc[1000])

def capture_conf_TK(x):
    # check index
    index = x.ix[0]
    if x['wox_info']:
        return TK_wox.iloc[index]['confidence']
    else:
        return TK_nowox.iloc[index]['confidence']
print(capture_conf_ee(predictions.iloc[1000]))
print(capture_conf_TK(predictions.iloc[1000]))

predictions['TK_conf']  = predictions.apply(lambda x: capture_conf_TK(x), axis = 1)
predictions['EE_conf']  = predictions.apply(lambda x: capture_conf_ee(x), axis = 1)

predictions.head()

#ee_wox_expl = pd.read_csv('../interpretations/')

# Capturing top 5 explanations

def capture_expl_ee(x):
    # check index
    index = x.ix[0]
    if x['wox_info']:
        # take top 5 from the dataframe
        temp = ee_wox.iloc[index].sort_values(ascending=False).drop(['Unnamed: 0', 'confidence'])[:5]
        temp_index, temp_values = [x for x in temp.index], [x for x in temp.values]
        # we zip so we can iterate
        return [f'{item[0]}-{item[1]:.1f}' for item in zip(temp_index,temp_values)]
    else:
        temp = ee_nowox.iloc[index].sort_values(ascending=False).drop(['Unnamed: 0', 'confidence'])[:5]
        temp_index, temp_values = [x for x in temp.index], [x for x in temp.values]
        return [f'{item[0]}-{item[1]:.1f}' for item in zip(temp_index,temp_values)]

def capture_expl_tk(x):
    # check index
    index = x.ix[0]
    if x['wox_info']:
        return TK_wox.iloc[index].sort_values(ascending=False).drop(['Unnamed: 0', 'confidence']).index[:5]
    else:
        return TK_nowox.iloc[index].sort_values(ascending=False).drop(['Unnamed: 0', 'confidence']).index[:5]
print(capture_expl_ee(predictions.iloc[1000]))
print(capture_expl_tk(predictions.iloc[1000]))

temp = ee_wox.iloc[1000].sort_values(ascending=False).drop(['Unnamed: 0', 'confidence'])[:5]

temp_index = [x for x in temp.index]
temp_values = [x for x in temp.values]

temp.values

zipped = [x for x in zip(temp_index,temp_values)]

feature_list = []
for item in zipped:
    to_add = f'{item[0]}-{item[1]:.1f}'
    feature_list.append(to_add)

[f'{item[0]}-{item[1]:.1f}' for item in zip(temp_index,temp_values)]
